Changed getting / setting of default rounding mode.
With a previous commit the current default rounding mode has been stored in the
enum class ROUNDING itself. This is voked now. Instead, the function's g'get_rounding'
and 'set_rounding' are moved to the module '_pydecimalfp' and the current default
rounding mode is stored as a (protected) variable in that module.